Fix #438 Handle asset deletion (#439)
in this current implementation, we will sweep the asset/price
immediately. it's the caller duty to cleanup all the tasks. In our
current mode, we can only run this through government so we trust the
caller to do the task cleanup before running the asset removal

currently we use a feature flag to delete asset, but once we enable
sudo we can use it and remove the flag.